requirements.txt
# Core dependencies
fastapi
uvicorn[standard]
python-multipart
jinja2
boto3
python-dotenv

# Testing dependencies
pytest
pytest-asyncio
httpx